optimize build.sh

+++ build.sh
@@ -1,40 +1,45 @@
 #!/bin/bash
+
 basedir=`pwd`
 
 function build_vim(){
-  ln -s $basedir/vimfiles/vim ~/.vim
-  ln -s $basedir/vimfiles/vimrc ~/.vimrc
+  [[ ! -d ~/.vim ]] && ln -s $basedir/vimfiles/vim ~/.vim
+  [[ ! -f ~/.vimrc ]] && ln -s $basedir/vimfiles/vimrc ~/.vimrc
 
-  mkdir -p ~/.vim/bundle
-  git clone http://github.com/gmarik/vundle.git ~/.vim/bundle/vundle
+  [[ ! -d ~/.vim/bundle ]] && mkdir -p ~/.vim/bundle
+  [[ ! -d ~/.vim/bundle/vundle ]] && git clone http://github.com/gmarik/vundle.git ~/.vim/bundle/vundle
   vim +BundleInstall +qa
   git submodule init; 
   git submodule update
 
-  if [[ `which rvm` ]]
+  if [[ `which rvm` &&  -d "~/.vim/bundle/Command-T" ]]
   then
-      if [[ -d "~/.vim/bundle/Command-T" ]]
-       cd ~/.vim/bundle/Command-T/ruby/command-t/
-       rvm system do ruby extconf.rb
-       make
-       cd - 
-   fi
+      cd ~/.vim/bundle/Command-T/ruby/command-t/
+      rvm system do ruby extconf.rb
+      make
+      cd - 
+  fi
 }
 
 function build_bash(){
    echo 'build bash' 
-   ln -s $basedir/bash/bash_profile ~/.bash_profile
-   ln -s $basedir/bash/bash_alias ~/.bash_alias
-   ln -s $basedir/bash/markrc ~/.markrc
+   echo "create links"
+   [[ ! -f ~/.bash_profile ]] && ln -s $basedir/bash/bash_profile ~/.bash_profile
+   [[ ! -f ~/.bash_alias ]] && ln -s $basedir/bash/bash_alias ~/.bash_alias
+   [[ ! -f ~/.markrc ]] && ln -s $basedir/bash/markrc ~/.markrc
+   [[ ! -f ~/.zshrc ]] && ln -s $basedir/bash/zshrc ~/.zshrc
+   [[ ! -f ~/.screenrc ]] && ln -s $basedir/screen/screenrc ~/.screenrc
+   echo "install zsh"
    # install zsh
    cd ~
-   git clone https://github.com/robbyrussell/oh-my-zsh.git 
-   ln -s $basedir/bash/zshrc ~/.zshrc
-   ln -s $basedir/screen/screenrc ~/.screenrc
-   mkdir ~/bin
-   cp $basedir/ssh.exp ~/bin/
-   cp $basedir/*.sh ~/bin/
-   chown +x ~/bin/*.sh ~/bin/*.exp
+   if [[ ! -d .oh-my-zsh ]]
+   then
+       git clone https://github.com/robbyrussell/oh-my-zsh.git .oh-my-zsh
+   fi
+   [[ ! -d ~/bin ]] && mkdir ~/bin
+   cp $basedir/bash/*.exp ~/bin/
+   cp $basedir/bash/*.sh ~/bin/
+   chmod +x ~/bin/*.sh ~/bin/*.exp
 }
